requirements.txt
streamlit==1.31.0
pandas==2.2.0
streamlit-aggrid==0.3.4
plotly==5.18.0
supabase==2.3.4
python-dotenv==1.0.1
groq==0.4.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
pymupdf==1.23.21
pydantic==2.6.1

